Make sure `gdkconfig.h' exists after the rule is fired.
* gdk/Makefile.am (gdkconfig.h): Make sure `gdkconfig.h' exists
after the rule is fired.
(install-exec-local): Install gdkconfig.h only if the contents are
different from the currently installed gdkconfig.h.

Rationale:
`make install-exec' is meant to install files that go into
$exec_prefix, which includes `gdkconfig.h'.  However, developers use
`make install-exec' when they know that none of the headers have
changed, to prevent needless recompiles of other packages.  My
previous change of installing `*config.h' during `make install-exec'
undermines this, even though it is the right thing to do.  So, in the
spirit of "do the right thing first, optimize later", here's the
optimization.