Implemented virtual atoms for gdk/gtk
changed all the gtk multihead design remove screen variable from GtkWidget to GtkWindow
now gtk_widget_get_screen recurse through the widget tree to find a top level window
with the screen information (if not found check special case e.g. GtkInvisible or
screen info attached to a widget).

Rewrote GtkSelection multihead abstraction to reflect virtual atom changes and the
fact that the display information is not available a destruction time. Now it is buggy ;(

Rewrote GtkClipboard multihead abstraction to remove dependency on the display information
as it is not available at destruction time.

All the atoms are virtual in gtk, in gdk atoms are translated into virtual atoms as the
GdkEvents are constructed (the only remaining real atoms are WM related atoms and
Motif Dnd target table).