Remove a non-effective style-set handler
The style-set handler was supposed to set the background color of
the text views in the credits dialog, but due to a thinko it never
had any effect, unless a theme change happened while the dialog
was open. Seems best to keep the current appearance now and remove
the handler. (#577868)